+++ app/services/schedule_generator.py
@@ -1,5 +1,5 @@
-from app.models.week import Week
-from app.models.subject_count import SubjectCount
+from week import Week
+from subject_count import SubjectCount
 
 class ScheduleGenerator:
